More files to C++.
Switch a few more files to C++ compilation and resolve issues found: add
several extern "C" declarations, remove unused variables, move some
static variables from headers to source files.  Move gmx_ctime_r()
declaration to string2.h; it must be better to have one extra
declaration in an installed header than a mass of repeated declarations
in source files.  Remove a hack in smalloc.h; the code that prompted
that seems to have been removed in the kernel reorganization for 4.6,
and we are anyways not yet compiling a lot of legacy code with C++.
Suppress most scanf() warnings in these files.  Make cppcheck correctly
parse #ifdef __cplusplus blocks.

The motivation for this selection of files is that these files write
output files where either command_line() or Program() is used to
indicate the currently running program.  Will replace those calls with a
uniform mechanism in a subsequent commit.

+++ src/gromacs/legacyheaders/typedefs.h
@@ -99,10 +99,10 @@ int over_alloc_dd(int n);
  */
 
 /* Over allocation for small data types: int, real etc. */
-#define over_alloc_small(n) (OVER_ALLOC_FAC*(n) + 8000)
+#define over_alloc_small(n) (int)(OVER_ALLOC_FAC*(n) + 8000)
 
 /* Over allocation for large data types: complex structs */
-#define over_alloc_large(n) (OVER_ALLOC_FAC*(n) + 1000)
+#define over_alloc_large(n) (int)(OVER_ALLOC_FAC*(n) + 1000)
 
 int gmx_large_int_to_int(gmx_large_int_t step, const char *warn);
 /* Convert a gmx_large_int_t value to int.
